scrapers/pyproject.toml
[project]
name = "grimdealz-scrapers"
version = "0.1.0"
description = "GrimDealz price scrapers — polls authorized US GW retailers every 4h"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27.0",
    "playwright>=1.44.0",
    "tenacity>=8.3.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.22.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",       # httpx mock
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
